+++ City.h
@@ -5,11 +5,11 @@ namespace Orienteering
 	class City
 	{
 	public:
-		int x;
-		int y;
+		double x;
+		double y;
 		int prize;
 		std::string c_name; 
-		City(int x=0, int y=0, int prize=-1, std::string name = "");
+		City(double x=0, double y=0, int prize=-1, std::string name = "");
 		bool operator==(City c);
 
 	};
